pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["mym1chelle <timur.a.samusenko@mail.ru>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.8.14"
flask = "^2.2.2"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.0"
importlib-metadata = "^5.1.0"
psycopg2-binary = "^2.9.5"
validators = "^0.20.0"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
